fix data type

+++ mros2-mbed/workspace/proj_node/Node_MotorControl/app.cpp
@@ -24,7 +24,7 @@
 #include <string>
 
 void splitData(std::string cmData);
-void frontControl(std::string frontDirection, int diff_degree);
+void frontControl(std::string frontDirection, float diff_degree);
 void motorControl(int period_PWM, float dutycycle_PWM, std::string backDirection);
 
 DigitalIn signalPinR(PF_12);
@@ -88,7 +88,7 @@ void splitData(std::string cmData)
   if(std::getline(ss, token, ',')){
     backDirection = token;
   }
-  
+
   frontControl(frontDirection, frontDegree);
   motorControl(period_PWM, dutycycle_PWM, backDirection);
 }
